fix: use weights.sum() as effective n in IC and tau for sample_weight equivalence
- abessLm/abessMulGaussian loss_function: divide by weights.sum() not n
- Algorithm::update_tau: change signature to double, pass train_weight.sum()
- AlgorithmPCA/Ising/Graph::update_tau: update override signatures to double
- Metric::ic: change train_n parameter to double, use data.weight.sum()
- workflow.h/path.h: pass data.weight.sum()/train_weight.sum() to ic()

This ensures WLS (sample_weight) and repeated-data give identical IC values,
active set selection, and model coefficients. All 14 Python tests pass.

via [Happy](https://happy.engineering)

+++ src/AlgorithmGraph.h
@@ -18,7 +18,7 @@ public:
 
   ~abessGraph(){};
 
-  void update_tau(int train_n, int N)
+  void update_tau(double train_n, int N)
   {
     // cout<<"init tau"<<endl;
     this->tau = 0.01 * (double)this->sparsity_level * log((double)N - (this->x).cols()) * log(log((double)train_n)) / (double)train_n;
