+++ src/AlgorithmPCA.h
@@ -139,7 +139,7 @@ class abessRPCA : public Algorithm<Eigen::VectorXd, Eigen::VectorXd, double, T4>
 
     int get_beta_size(int n, int p) { return n * p; }
 
-    void update_tau(int train_n, int N) { this->tau = 0.0; }
+    void update_tau(double train_n, int N) { this->tau = 0.0; }
 
     Eigen::VectorXi inital_screening(T4 &X, Eigen::VectorXd &y, Eigen::VectorXd &beta, double &coef0,
                                      Eigen::VectorXi &A, Eigen::VectorXi &I, Eigen::VectorXd &bd,
